[REFACTOR] Remove tvm.runtime.object shim and route Object via tvm_ffi
Drop python/tvm/runtime/object.py and route tvm.runtime.Object directly
to tvm_ffi.Object. The load-bearing __object_repr__ install (which wires
TVM IR objects up to the rich C++ ReprPrinter registered through
init_ffi_api("node", ...)) moves into runtime/_ffi_node_api.py, which
is already imported as a side-effect-only module from runtime/__init__.py.

The remaining importers of the old shim are switched to
`from tvm.runtime import Object` (or `from tvm_ffi import Object` where
that matches the file's existing tvm_ffi imports).
