--- python/tvm/runtime/object.py
@@ -1,27 +1,0 @@
-# Licensed to the Apache Software Foundation (ASF) under one
-# or more contributor license agreements.  See the NOTICE file
-# distributed with this work for additional information
-# regarding copyright ownership.  The ASF licenses this file
-# to you under the Apache License, Version 2.0 (the
-# "License"); you may not use this file except in compliance
-# with the License.  You may obtain a copy of the License at
-#
-#   http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing,
-# software distributed under the License is distributed on an
-# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-# KIND, either express or implied.  See the License for the
-# specific language governing permissions and limitations
-# under the License.
-# pylint: disable=invalid-name, unused-import
-"""Runtime Object API"""
-
-import tvm_ffi.core
-from tvm_ffi.core import Object
-
-from . import _ffi_node_api
-
-tvm_ffi.core._set_class_object(Object)
-# override the default repr function for tvm_ffi.core.Object
-tvm_ffi.core.__object_repr__ = _ffi_node_api.AsRepr
